resource_server/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
httpx
pydantic
